// src/models/Accommodation.ts
export interface Accommodation {
    id: string;
    title: string;
    cover: string;
    pictures: string[];
    description: string;
    host: Host;
    rating: string;
    location: string;
    equipments: string[];
    tags: string[];
}

export interface Host {
    name: string;
    picture: string;
}